pyproject.toml
[tool.poetry]
name = "metalbot"
version = "0.1.0"
description = ""
authors = ["workingakk <autokat.k@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
bs4 = "^0.0.2"
aiogram = "^3.4.1"
python-dotenv = "^1.0.1"
schedule = "^1.2.1"
pandas = "^2.2.1"
openpyxl = "^3.1.2"
pillow = "^10.3.0"
xlrd = "^2.0.1"
asyncpg = "^0.29.0"


[tool.poetry.group.lint.dependencies]
black = "^24.3.0"
flake8 = "^7.0.0"
isort = "^5.13.2"
mypy = "^1.9.0"

[tool.black]
line-length = 119

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = [".venv", "tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
